+++ artscription-ui/src/app/artscriptions/page.tsx
@@ -1,9 +1,8 @@
+import Artscription from "@/ui/artscriptions";
 export default async function Page() {
     return (
         <main>
-            <h1 className={` mb-4 text-xl md:text-2xl`}>
-                artscriptions
-            </h1>
+            <Artscription />
         </main>
     );
 }
