+++ artscription-ui/src/app/marketplace/page.tsx
@@ -1,9 +1,10 @@
+import Marketplace from "@/ui/marketplace";
 export default async function Page() {
     return (
         <main>
-            <h1 className={` mb-4 text-xl md:text-2xl`}>
-                marketplace
-            </h1>
+       
+            <Marketplace/>
+            
         </main>
     );
 }
