+++ artscription-ui/src/ui/artscriptions.tsx
@@ -1,5 +1,5 @@
-// pages/artscription.js
-import '../app/globals.css';
+"use client"
+import './globals.css';
 import React, { useEffect, useState } from 'react';
 
 export default function Artscription() {
