+++ artscription-ui/src/ui/lib/artscriptionDetail.js
